split out ElementalInertialMatrices
